refactor(matcher): extract base methods to the abstract class
