+++ src/providers/matchers/partial.spec.ts
@@ -1,8 +1,7 @@
 import * as assert from 'assert';
 
-import * as tests from '../../tests';
 import { Pattern, MicromatchOptions } from '../../types';
-import Matcher, { PatternInfo } from './partial';
+import Matcher from './partial';
 
 function getMatcher(patterns: Pattern[], options: MicromatchOptions = {}): Matcher {
 	return new Matcher(patterns, options);
@@ -21,26 +20,6 @@ function assertNotMatch(patterns: Pattern[], level: number, part: string): void 
 }
 
 describe('Providers → Matchers → Partial', () => {
-	describe('.storage', () => {
-		it('should return created storage', () => {
-			const matcher = getMatcher(['a*', 'a/**/b']);
-
-			const expected: PatternInfo[] = [
-				tests.pattern.info()
-					.section(tests.pattern.segment().dynamic().pattern('a*').build())
-					.build(),
-				tests.pattern.info()
-					.section(tests.pattern.segment().pattern('a').build())
-					.section(tests.pattern.segment().pattern('b').build())
-					.build()
-			];
-
-			const actual = matcher.storage;
-
-			assert.deepStrictEqual(actual, expected);
-		});
-	});
-
 	describe('.match', () => {
 		it('should handle patterns with globstar', () => {
 			assertMatch(['**'], 0, 'a');
