+++ src/providers/matchers/partial.ts
@@ -1,42 +1,6 @@
-import { Pattern, PatternRe, MicromatchOptions } from '../../types';
-import * as utils from '../../utils';
+import Matcher from './matcher';
 
-export type PatternSection = PatternSegment[];
-
-export type PatternSegment = StaticPatternSegment | DynamicPatternSegment;
-
-type StaticPatternSegment = {
-	dynamic: false;
-	pattern: Pattern;
-};
-
-type DynamicPatternSegment = {
-	dynamic: true;
-	pattern: Pattern;
-	patternRe: PatternRe;
-};
-
-export type PatternInfo = {
-	/**
-	 * Indicates that the pattern has a globstar (more than a single section).
-	 */
-	complete: boolean;
-	pattern: Pattern;
-	segments: PatternSegment[];
-	sections: PatternSection[];
-};
-
-export default class PartialMatcher {
-	private readonly _storage: PatternInfo[] = [];
-
-	constructor(private readonly _patterns: Pattern[], private readonly _options: MicromatchOptions) {
-		this._fillStorage();
-	}
-
-	public get storage(): PatternInfo[] {
-		return this._storage;
-	}
-
+export default class PartialMatcher extends Matcher {
 	public match(level: number, part: string): boolean {
 		for (const info of this._storage) {
 			const section = info.sections[0];
@@ -75,29 +39,5 @@ export default class PartialMatcher {
 		}
 
 		return false;
-	}
-
-	private _fillStorage(): void {
-		/**
-		 * The original pattern may include `{,*,**,a/*}`, which will lead to problems with matching (unresolved level).
-		 * So, before expand patterns with brace expansion into separated patterns.
-		 */
-		const patterns = utils.pattern.expandPatternsWithBraceExpansion(this._patterns);
-
-		for (const pattern of patterns) {
-			const segments = utils.pattern.getPatternSegments(pattern, this._options);
-			const sections = this._splitSegmentsIntoSections(segments);
-
-			this._storage.push({
-				complete: sections.length <= 1,
-				pattern,
-				segments,
-				sections
-			});
-		}
-	}
-
-	private _splitSegmentsIntoSections(segments: PatternSegment[]): PatternSection[] {
-		return utils.array.splitWhen(segments, (segment) => segment.dynamic && utils.pattern.hasGlobStar(segment.pattern));
 	}
 }
