+++ src/utils/pattern.spec.ts
@@ -1,7 +1,6 @@
 import * as assert from 'assert';
 
-import { Pattern, PatternSegment } from '../types';
-import * as tests from '../tests';
+import { Pattern } from '../types';
 import * as util from './pattern';
 
 describe('Utils → Pattern', () => {
@@ -381,22 +380,6 @@ describe('Utils → Pattern', () => {
 			const expected: Pattern[] = ['a', '*', 'b', '**', 'c'];
 
 			const actual = util.getPatternParts('a/*/b/**/c', {});
-
-			assert.deepStrictEqual(actual, expected);
-		});
-	});
-
-	describe('.getPatternSegments', () => {
-		it('should return an array of pattern segments', () => {
-			const expected: PatternSegment[] = [
-				tests.pattern.segment().pattern('a').build(),
-				tests.pattern.segment().dynamic().pattern('*').build(),
-				tests.pattern.segment().pattern('b').build(),
-				tests.pattern.segment().dynamic().pattern('**').build(),
-				tests.pattern.segment().pattern('c').build()
-			];
-
-			const actual = util.getPatternSegments('a/*/b/**/c', {});
 
 			assert.deepStrictEqual(actual, expected);
 		});
